--- apps/person-export-api/src/main/java/no/nav/registre/testnorge/personexportapi/PersonExportApiApplicationStarter.java
@@ -1,11 +1,0 @@
-package no.nav.registre.testnorge.personexportapi;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class PersonExportApiApplicationStarter {
-    public static void main(String[] args) {
-        SpringApplication.run(PersonExportApiApplicationStarter.class, args);
-    }
-}
