+++ apps/person-export-api/src/main/java/no/nav/testnav/apps/personexportapi/PersonExportApiApplicationStarter.java
@@ -1,0 +1,11 @@
+package no.nav.testnav.apps.personexportapi;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class PersonExportApiApplicationStarter {
+    public static void main(String[] args) {
+        SpringApplication.run(PersonExportApiApplicationStarter.class, args);
+    }
+}
