+++ apps/person-export-api/src/main/java/no/nav/testnav/apps/personexportapi/config/DevVaultConfig.java
@@ -1,0 +1,30 @@
+package no.nav.testnav.apps.personexportapi.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Profile;
+import org.springframework.vault.annotation.VaultPropertySource;
+import org.springframework.vault.authentication.ClientAuthentication;
+import org.springframework.vault.authentication.TokenAuthentication;
+import org.springframework.vault.client.VaultEndpoint;
+import org.springframework.vault.config.AbstractVaultConfiguration;
+
+@Configuration
+@Profile("dev")
+@VaultPropertySource(value = "azuread/prod/creds/team-dolly-lokal-app", ignoreSecretNotFound = false)
+@VaultPropertySource(value = "kv/preprod/fss/testnorge-person-export-api/dev", ignoreSecretNotFound = false)
+public class DevVaultConfig extends AbstractVaultConfiguration {
+
+    @Override
+    public VaultEndpoint vaultEndpoint() {
+        return VaultEndpoint.create("vault.adeo.no", 443);
+    }
+
+    @Override
+    public ClientAuthentication clientAuthentication() {
+        var token = System.getProperty("spring.cloud.vault.token");
+        if (token == null) {
+            throw new IllegalArgumentException("Påkreved property 'spring.cloud.vault.token' er ikke satt.");
+        }
+        return new TokenAuthentication(System.getProperty("spring.cloud.vault.token"));
+    }
+}
