+++ apps/person-export-api/src/main/java/no/nav/testnav/apps/personexportapi/config/SecurityConfig.java
@@ -1,0 +1,32 @@
+package no.nav.testnav.apps.personexportapi.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Profile;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
+import org.springframework.security.config.http.SessionCreationPolicy;
+
+@EnableWebSecurity
+@Configuration
+@Profile({"prod", "dev"})
+public class SecurityConfig extends WebSecurityConfigurerAdapter {
+
+    @Override
+    protected void configure(HttpSecurity http) throws Exception {
+        http.sessionManagement()
+                .sessionCreationPolicy(SessionCreationPolicy.STATELESS)
+                .and()
+                .csrf()
+                .disable()
+                .authorizeRequests()
+                .antMatchers("/api/**")
+                .fullyAuthenticated()
+                .and()
+                .oauth2ResourceServer()
+                .jwt();
+    }
+}
+
+
+
