--- apps/person-export-api/src/test/java/no/nav/registre/testnorge/personexportapi/ApplicationContextTest.java
@@ -1,20 +1,0 @@
-package no.nav.registre.testnorge.personexportapi;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.mock.mockito.MockBean;
-import org.springframework.security.oauth2.jwt.JwtDecoder;
-import org.springframework.test.context.ActiveProfiles;
-
-@SpringBootTest
-@ActiveProfiles("test")
-class ApplicationContextTest {
-
-    @MockBean
-    public JwtDecoder jwtDecoder;
-
-    @Test
-    @SuppressWarnings("java:S2699")
-    void load_app_context() {
-    }
-}
