+++ apps/person-export-api/src/test/java/no/nav/testnav/apps/personexportapi/ApplicationContextTest.java
@@ -1,0 +1,20 @@
+package no.nav.testnav.apps.personexportapi;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.test.mock.mockito.MockBean;
+import org.springframework.security.oauth2.jwt.JwtDecoder;
+import org.springframework.test.context.ActiveProfiles;
+
+@SpringBootTest
+@ActiveProfiles("test")
+class ApplicationContextTest {
+
+    @MockBean
+    public JwtDecoder jwtDecoder;
+
+    @Test
+    @SuppressWarnings("java:S2699")
+    void load_app_context() {
+    }
+}
